QCLINUX: virt: gunyah: Write KASLR addr to IMEM
Write KASLR address to IMEM. KASLR offset is required by various debug
tools for loading and parsing ramdumps for debug of crash issues.
Adding write of KASLR in init_watchdog of gh_virt_wdt. This is done
as a temp patch for 1.0. Writing this in gh_virt_wdt as this is mostly
the only file we want to carry currently as downstream in 1.0 to support
dload mode.

Change-Id: Ib7cf851ec87f679fc2197c7411574a5814a8aeb6
Signed-off-by: Shivendra Pratap <quic_spratap@quicinc.com>
(cherry picked from commit e0f17cedbc43a319d586d52d65959f6bb47eda44)
Signed-off-by: Hui Wang <hui.wang@canonical.com>