PENDING: usb: dwc3: Add vendor hook for run_stop notification
In case runtime autosuspend is enabled in device and UDC is removed,
the device goes to suspend. As part of suspend,  OTG_VBUS_VALID is
cleared and during re-writing of UDC, this bit is left cleared. Due
to this enumeration doesn't happen and once the autosuspend timer
expires, device enters suspend and enumeration fails completely.

Set the OTG_VBUS_VALID bit during run_stop(1) to prevent this
failure.

Change-Id: Ib419b658ba462312f1cc3949d5f64d72392531bd
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>