PENDING: Bluetooth: hci_qca: Drop unused event during BT on
For the WCN6750/WCN6855/WCN7850, the vendor command for a baudrate
change is not sent as synchronous HCI command, controller sends the
corresponding vendor event with the new baudrate. It needs to be
dropped, otherwise it may be misinterpreted as response to a later
command.

Change-Id: I29e6207cf3767adc2a7f8c5dd3e8a0d84a55be36
Signed-off-by: Cheng Jiang <quic_chejiang@quicinc.com>
(cherry picked from commit a01265f5cdb60e0ab0b8dc8fc595dc6e05a50f11)
Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>