PENDING: usb: dwc3: qcom: Introduce post-connection done vendor hook
During PIPO test cases on QCM6490, it has been sometimes observed
that upon cable connect, there is a role "0" coming in followed by
role "2" in success case. In fail case, there is no role "2" coming
in. But since the cable is connected at that instant and the device
is resumed from get_sync in cable disconnect hook, the qscratch
gets set from run stop vendor hook and enum is successful. However
since the role "2" is missing, the role is marked as NONE in software
of dwc3-qcom. When the next cable disconnect happens, the disconnect
notification is ignored by the cable disconnect vendor hook as dwc3
qcom is already in disconnected state according to it and hence it
doesn't clear the qscratch and suspend is blocked as disconnect
event won't be generated by the controller and the connected flag
in dwc structure is never set to false.

Prevent this by marking the role as DEVICE and cable disconnected
as false when connection done interrupt kicks in.

Change-Id: Ie47bd5879a386a7d1e191011cebb2b2fa0451772
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>
(cherry picked from commit 00b874383995152bc774e30c41dee612763d3b85)
Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>