QCLINUX: iommu: Turn on clocks before reading FSR
When fault occurs in smmu, the FSR register has to be clocked on
before reading the register. Otherwise, it results in NOC error
in the fault path.

Fix it by keeping the votes on for the register in fault path.

Change-Id: I661570fdb4099f3e349b3d69b00c4d28c088d3c1
Fixes: e93e2ace5aed ("BACKPORT: FROMLIST: iommu/arm-smmu-qcom: Use a custom context fault handler for sdm845")
Signed-off-by: Pratyush Brahma <quic_pbrahma@quicinc.com>
(cherry picked from commit 708a45116998f6bec53e9ee52e54a0b53f34de82
https://git.codelinaro.org/clo/la/kernel/qcom kernel.qclinux.1.0.r1-rel)
Signed-off-by: Hui Wang <hui.wang@canonical.com>