UBUNTU: SAUCE: usb: dwc3: make sure reset is deasserted before enabling clocks
BugLink: https://bugs.launchpad.net/carmel/bugs/2115323 ([RB3] ucsi_glink.pmic_glink_ucsi pmic_glink.ucsi.0: PPM init failed, stop trying)

When GCC clocks assigned to the dwc3 core are enabled, the halt_reg
register is used to verify if the clock changed state to the requested
one. If that's not the case, a WARN like this will be printed:

[    7.429459] gcc_usb30_prim_master_clk status stuck at 'off'

In addtion to that, an -EBUSY will be returned which causes an error
during dwc3 probe.

Problem is, the clock will not actually be enabled as long as the reset line is
asserted and this currently only happens inside of the dwc3_probe() (from
dwc3/core.c) which is called much later than the dwc3_qcom_clk_init()
which will trigger a problem described above. This problem can be
observed when the dwc3_qcom_probe is called 2nd time, for example after
returning EPROBE_DEFER on the first try.

To fix this problem, we have to make sure the reset is deasserted before
the dwc3_qcom_clk_init() is called. Note that since the same reset line will be
later also requested inside of the dwc3_probe(), the reset line needs to
be changed from exclusive to shared.

Signed-off-by: Krzysztof Adamski <krzysztof.adamski@canonical.com>
Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>