ASoC: qcom: qdsp6: q6apm: fix capture pipeline overruns
BugLink: https://bugs.launchpad.net/bugs/2117230 ([RB3][x04] alsa playback/record test failed)

Period sizes less than 6k for capture path triggers overruns in the
dsp capture pipeline.Change the period size and number of periods
to value which DSP need.

With the existing code, the buffer position is only reset in pointer
callback, which leaves the possiblity of it going over the size of
buffer size and reporting incorrect position to userspace.

DSP expects the periods to be aligned to fragment sizes, currently
setting up to hw constriants on periods bytes is not going to work
correctly as we can endup with periods sizes aligned to 32 bytes however
not aligned to fragment size.Update the constriants to use fragment size,
and also set at step of 10ms for period size to accommodate DSP
requirements of 10ms latency.

Change-Id: Ia44abe93352b9e4e79006678d24cb961daa434ef
Signed-off-by: Mohammad Rafi Shaik <quic_mohs@quicinc.com>
Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>