arm64: dts: qcom: switch from .mdt to .mbn firmwares
BugLink: https://bugs.launchpad.net/bugs/2121203 ([QC] Using mbn instead of mdt)

There are two firmware file formats supported by QC remoteproc - the
legacy one that is fragmented into .mdt and a banch of .bXX files and a
newer .mbn format that combines all that into one file. We currently
ship functionally same firmware in both formats and we want to move to
the newer and more convinient one.

This patch changes all the paths in DT that still uses .mdt format to
.mbn. This is relevant mostly for xDSP but also WPSS and modem are
affected.

The upstream (mainline) version of the qcs6490-rb3gen2.dts already uses
*.mbn files. The sa8775p-addons-ride.dts doesn't exist in mainline but
we want to also convert it so that we can drop shipping .mdt files
altogether.

Signed-off-by: Krzysztof Adamski <krzysztof.adamski@canonical.com>
Signed-off-by: Hui Wang <hui.wang@canonical.com>