PENDING: misc: fastrpc: Return on argument copy failure
If copying data from kernel to user space fails, the current code
still tries to call the DSP and release resources. But if unmapping
those resources fails, it can lead to resoursce leak and potential
fault issues.

Avoid this by not calling unmap and free in case of failure and rely
on process exit for cleanup.

Change-Id: I7baf60da186abb579ef5efdc7ff6626ec4f97713
Signed-off-by: Abhinav Parihar <quic_parihar@quicinc.com>
Signed-off-by: Hui Wang <hui.wang@canonical.com>