UBUNTU: ASoC: qcs9100: fix the driver name for RB8
BugLink: https://bugs.launchpad.net/bugs/2125094 ([RB4/RB8] wrong folder for topology file)

The driver_name is used as a base for the path to load the topology
files and UCM configs. The folder in /lib/firmware/ and in
/usr/share/alsa/ucm2/ that will be searched is defined by this field.
Since upstream (for historical reasons) is using "sa8775p", lets adopt
this convention as well. See:

https://github.com/alsa-project/alsa-ucm-conf/pull/618
https://web.git.kernel.org/pub/scm/linux/kernel/git/firmware/linux-firmware.git/commit/?id=6625e563d600eba34cf2c71df7572ad1c421c6fd

Signed-off-by: Krzysztof Adamski <krzysztof.adamski@canonical.com>
Signed-off-by: Hui Wang <hui.wang@canonical.com>